remove graphiz for tests on cibuildwheel
