+++ F2Py.cmake
@@ -184,12 +184,6 @@ function (f2py_add_module target_name)
     target_link_libraries(${target_name} PUBLIC "-static")
   endif()
 
-  if (APPLE)
-    set_target_properties(${target_name} PROPERTIES
-      BUILD_RPATH "/usr/local/lib"
-      INSTALL_RPATH "/usr/local/lib"    
-    )
-  endif()
 
   set_property(TARGET ${target_name} PROPERTY SUFFIX ${PYTHON_MODULE_EXTENSION})
   # must be a string, so that empty string works correcty
